add base_extractor to galaxyzoo module, use class label keys and prefexises for extractor tests, remove error on baseclass
