+++ app/modules/label_extractors/galaxy_zoo/decals.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require_relative '../base_extractor'
 
 module LabelExtractors
   module GalaxyZoo
