+++ lib/bajor/client.rb
@@ -25,7 +25,7 @@ module Bajor
         '/training/jobs/',
         # NOTE: Here we can augment the batch job run options via bajor
         # via the `run_opts: '--wandb --debug'` etc, these could be set via ENV
-        body: { manifest_path: manifest_path, opts: { 'run_opts': "--schema #{workflow_name}", 'workflow_name': workflow_name } }.to_json,
+        body: { manifest_path: manifest_path, opts: { 'run_opts': "--schema #{workflow_name.downcase}", 'workflow_name': workflow_name } }.to_json,
         headers: JSON_HEADERS
       )
