+++ spec/modules/label_extractors/galaxy_zoo/base_extractor_spec.rb
@@ -1,6 +1,6 @@
 # spec/models/label_extractors/base_extractor_spec.rb
 
-class TestExtractor < LabelExtractors::BaseExtractor
+class TestExtractor < LabelExtractors::GalaxyZoo::BaseExtractor
   TASK_KEY_LABEL_PREFIXES = { task_key_example: "example_prefix" }
   TASK_KEY_DATA_LABELS = { task_key_example: { "T0" => "example_label" } }
   def self.data_release_suffix
@@ -8,7 +8,7 @@ class TestExtractor < LabelExtractors::BaseExtractor
   end
 end
 
-RSpec.describe LabelExtractors::BaseExtractor do
+RSpec.describe LabelExtractors::GalaxyZoo::BaseExtractor do
   let(:task_lookup_key) { :task_key_example }
   let(:data_hash) { { "T0" => "example_key" } }
   let(:instance) { TestExtractor.new(task_lookup_key) }
@@ -35,7 +35,7 @@ RSpec.describe LabelExtractors::BaseExtractor do
     it 'raises an error for an unknown task_lookup_key' do
       invalid_instance = TestExtractor.allocate # Skips calling initialize
       allow(invalid_instance).to receive(:task_lookup_key).and_return(:invalid_key)
-      expect { invalid_instance.send(:task_prefix) }.to raise_error(LabelExtractors::BaseExtractor::UnknownTaskKey)
+      expect { invalid_instance.send(:task_prefix) }.to raise_error(LabelExtractors::GalaxyZoo::UnknownTaskKey)
     end
   end
 
@@ -45,7 +45,7 @@ RSpec.describe LabelExtractors::BaseExtractor do
     end
 
     it 'raises an error for an unknown key' do
-      expect { instance.send(:data_payload_label, "unknown_key") }.to raise_error(LabelExtractors::BaseExtractor::UnknownLabelKey)
+      expect { instance.send(:data_payload_label, "unknown_key") }.to raise_error(LabelExtractors::GalaxyZoo::UnknownLabelKey)
     end
   end
 end
